hotfix: size_t typing for memory functions

+++ srcs/int/ft_dup_int_arr.c
@@ -6,17 +6,17 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/18 23:08:41 by dcastor           #+#    #+#             */
-/*   Updated: 2025/05/22 11:58:42 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/05/22 14:38:05 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-int	*ft_dup_int_arr(int arr[], unsigned int size)
+int	*ft_dup_int_arr(int arr[], size_t size)
 {
-	int				*new_arr;
-	unsigned int	i;
+	int		*new_arr;
+	size_t	i;
 
 	new_arr = malloc(sizeof(int) * size);
 	if (!new_arr)
