+++ srcs/int/ft_int_in_arr.c
@@ -6,15 +6,15 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/19 11:06:52 by dcastor           #+#    #+#             */
-/*   Updated: 2025/05/22 11:58:53 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/05/22 14:38:16 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_int_in_array(int *arr, unsigned int size, int value)
+int	ft_int_in_array(int *arr, size_t size, int value)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < size)
